*Added new ehcmodules with better and more drives support on Hermes IOS (thx rodries) *Fixed reset of the loader when loading game with IOS reload and disabled WiiTDB titles (Issue 1874) *Added new 'Inherit' or 'Use global' setting for game settings. If that option is set than the global setting is used for that option. (Issue 1842) *Fixed timeout timer on startup to count correctly. (Issue 1907) *Added two new video modes to force progressive video mode, 'FORCE PAL480p' and 'FORCE NTSC480p' (basically the same but oh well) (Issue 1902) *Added the new 'Return to' procedure for the d2x v4 IOS for Test purpose (didn't test it, need feedback on this one). The old method is used if this procedure fails. Please test it on problematic games (e.g. PoP). (Issue 1892)

+++ source/settings/CGameSettings.h
@@ -10,20 +10,20 @@
 typedef struct _GameCFG
 {
     char id[7];
-    u8 video;
-    u8 language;
-    u8 ocarina;
-    u8 vipatch;
-    u8 ios;
-    u8 parentalcontrol;
-    u8 errorfix002;
-    u8 iosreloadblock;
-    u8 loadalternatedol;
+    short video;
+    short language;
+    short ocarina;
+    short vipatch;
+    short ios;
+    short parentalcontrol;
+    short errorfix002;
+    short iosreloadblock;
+    short loadalternatedol;
     u32 alternatedolstart;
-    u8 patchcountrystrings;
+    short patchcountrystrings;
     char alternatedolname[40];
-    u8 returnTo;
-    u8 Locked;
+    short returnTo;
+    short Locked;
 } GameCFG;
 
 class CGameSettings
